% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_dependency_to_fields.R
\name{add_dependency_to_fields}
\alias{add_dependency_to_fields}
\title{Add an additional dependency to certain columns}
\usage{
add_dependency_to_fields(
  these_dependencies,
  field_match,
  new_dependency,
  verbose = FALSE
)
}
\arguments{
\item{these_dependencies}{A dependency map}

\item{field_match}{A string pattern that will identify the columns to update}

\item{new_dependency}{The new dependency to add}

\item{verbose}{Whether informative messages should be printed (default = FALSE)}
}
\value{
An updated dependency map.
}
\description{
This function takes a dependency map (these_dependencies), looks
for fields #' that match a particular string pattern (field_match)
and adds a new dependency (new_dependency) to the column for
manual dependencies (param_dependencies_manual).
If there is nothing there, the new dependency becomes the whole
dependency.  Any columns that are updated will be marked as
amended in the dependency map tibble.
}
