% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_files_exist.R
\name{all_files_exist}
\alias{all_files_exist}
\title{Do all the files listed actually exist?}
\usage{
all_files_exist(file_list)
}
\arguments{
\item{file_list}{A list of files}
}
\value{
TRUE/FALSE do all files exist?
}
\description{
Do all the files listed actually exist?
}
