% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_BMI.R
\name{calculate_BMI}
\alias{calculate_BMI}
\title{Given weight and height, returns BMI}
\usage{
calculate_BMI(weight, height)
}
\arguments{
\item{weight}{Weight, in kg.}

\item{height}{Height, in cm.}
}
\value{
Calculated BMI
}
\description{
Given weight and height, returns BMI
}
\examples{
calculate_BMI( 52, 155 )
}
