% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_dose_modification.R
\name{calculate_dose_modification}
\alias{calculate_dose_modification}
\title{Modify dose according to protocol schedule}
\usage{
calculate_dose_modification(
  value,
  direction = "+",
  dose_scale = TRIAL.dose_scale
)
}
\arguments{
\item{value}{The current value.}

\item{direction}{The direction (+/-) to move in on the dose scale (default="+")}

\item{dose_scale}{The existing dose escalation/descalation scale for this trial.}
}
\value{
The new dose.
}
\description{
This function takes an existing scale for dose modifications
(as TRIAL.dose_scale) and, given an input value and a direction
(+ for escalation and - for descalation), returns the modified dose.
}
\examples{
this.dose_scale = c( 1,2,4,8,16 )
calculate_dose_modification( 2 , dose_scale = this.dose_scale )
calculate_dose_modification( 16, "-", dose_scale = this.dose_scale )
}
