% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_missing_data.R
\name{calculate_missing_data}
\alias{calculate_missing_data}
\title{CAlculate how much data is missing for each patient, visit and test battery.}
\usage{
calculate_missing_data(
  this_data,
  this_glossary,
  id_var,
  site_var,
  specification
)
}
\arguments{
\item{this_data}{The tibble representing the patient population to be checked for missing data.}

\item{this_glossary}{The glossary dictionary with parameter visit/test battery etc information.}

\item{id_var}{A quosure representing the column that contains the patient label.}

\item{site_var}{A quosure representing the column that contains the site name for that patient.}

\item{specification}{Which fields for which patients should be considered when calculating how much data is missing.}
}
\value{
Missing data percentages in a tibble
}
\description{
CAlculate how much data is missing for each patient, visit and test battery.
}
