% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_number_of_instances.R
\name{calculate_number_of_instances}
\alias{calculate_number_of_instances}
\title{Calculate number of instances}
\usage{
calculate_number_of_instances(
  this_data,
  this_var,
  patient_column = NULL,
  timepoint_column = NULL,
  spread_timepoints = TRUE,
  this_function = NA
)
}
\arguments{
\item{this_data}{The tibble in which the column exists}

\item{this_var}{The column name (if not a quosure, it will be made into one)}

\item{patient_column}{The column name containing the patient ID to group by (default is to not group)}

\item{timepoint_column}{The column name containing the timepoint to group by (default is to not group)}

\item{spread_timepoints}{A column to spread the results by timepoint (default is TRUE)}

\item{this_function}{A label for error/warning messages (by default it is the name of this function).}
}
\value{
tibble containing the patient and their tumour counts for each timepoint.
}
\description{
This function returns the number of non-empty entries,
optionally for each patient and each timepoint.
Ensure that the tibble provided includes ALL rows (i.e.,
this is the extended form of the data from MACRO, with lots
of empty values).  If grouping by timepoint is specified,
the data will be returned with the timepoints spread as columns.
}
