% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_RECIST_response.R
\name{calculate_RECIST_response}
\alias{calculate_RECIST_response}
\title{Calculates whether data indicates a complete response by RECIST criteria}
\usage{
calculate_RECIST_response(
  this_data,
  id_variable,
  baseline_diam,
  baseline_sum,
  min_sum = NULL,
  timepoint_diam,
  timepoint_sum,
  this_function = NA,
  label_string = "RECIST"
)
}
\arguments{
\item{this_data}{The tibble in which the column exists.}

\item{id_variable}{The column that defines each patient (i.e., the patient ID)}

\item{baseline_diam}{The variable that represents the individual diameters at baseline}

\item{baseline_sum}{The variable that represents the sum at baseline}

\item{min_sum}{The variable that represents the smallest diameter in the study (if absent, assumes this value is the baseline_sum)}

\item{timepoint_diam}{The variable that represents the individual diameters at the later timepoint}

\item{timepoint_sum}{The variable that represents the sum at the later timepoint}

\item{this_function}{A label for error/warning messages (by default it is the name of this function)}

\item{label_string}{The string to use for the result column}
}
\value{
A list, containing (1) the assessment (using RECIST v1.1 criteria) and (2) the supporting data
}
\description{
This function takes tumour measurement data and calculates the RECIST outcome.
It calculates a response for one timepoint measurement against the baseline.
If multiple timepoint measurements are to be compared to a baseline, then this
function will need to be run multiple times.
It presumes that the data that are passed here are clean (i.e., all data are present,
none are missing; all diameters add up to the sum etc.).  The data in
should also have a column that represents the minimum sum of tumour diameters
(across all timepoints) for each patient.
}
