% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_all_files_exist.R
\name{check_all_files_exist}
\alias{check_all_files_exist}
\title{Check whether all files in a list exist}
\usage{
check_all_files_exist(file_list)
}
\arguments{
\item{file_list}{A list of files (will assume local directory if no directory provided)}
}
\value{
TRUE/FALSE to indicate whether all files exist
}
\description{
Check whether all files in a list exist
}
