% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_extract_date.R
\name{check_extract_date}
\alias{check_extract_date}
\title{Check whether the date recorded for the extract matches the file name.}
\usage{
check_extract_date(filename, trial, date_provided)
}
\arguments{
\item{filename}{The name of the extract file}

\item{trial}{The name of the trial}

\item{date_provided}{The recorded date of the extract}
}
\value{
Whether the name of the file and the extract date match
}
\description{
Check whether the date recorded for the extract matches the file name.
}
\examples{
check_extract_date( "NAXIVA20200304.csv", "NAXIVA", lubridate::ymd("2020-03-04") )
check_extract_date( "NAXIVA20200304.csv", "NAXIVA", lubridate::ymd("2022-03-04")  )
}
