% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EUDRACT_functions.R
\name{convert_dates_to_EudraCT_age_categories}
\alias{convert_dates_to_EudraCT_age_categories}
\title{Given Date of Birth and Date of Registration, annotate the
patient with their EudraCT factory.}
\usage{
convert_dates_to_EudraCT_age_categories(dob, dor)
}
\arguments{
\item{dob}{A list of Dates of birth (as lubridate objects)}

\item{dor}{A list of Dates of registration (as lubridate objects)}
}
\value{
A list of EudraCT age catetgories
}
\description{
Given Date of Birth and Date of Registration, annotate the
patient with their EudraCT factory.
}
\examples{
convert_dates_to_EudraCT_age_categories(
     lubridate::ymd( "1982-04-05" ),
     lubridate::ymd( "2020-10-06" ) )
}
