% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dummy_data.R
\name{create_dummy_dataset}
\alias{create_dummy_dataset}
\title{Create dummy CT data}
\usage{
create_dummy_dataset(
  number_of_patients = 100,
  perc_missing_data = 0.05,
  use_seed = NA
)
}
\arguments{
\item{number_of_patients}{The number of patients in the dataset (default=100)}

\item{perc_missing_data}{What percentage of data should be missing (default=0.05)}

\item{use_seed}{A seed to use (default=NA, i.e., no specified seed)}
}
\value{
Data objects for dummy dataset as a list: (1) the data, (2) the vocabulary, (3) the glossary.
The vocabulary tells you what values each column can take.
The glossary gives you a display string for each column.
}
\description{
Create dummy CT data
}
