% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_SSR_structure.R
\name{create_SSR_structure}
\alias{create_SSR_structure}
\title{Create a new SSR directory}
\usage{
create_SSR_structure(template)
}
\arguments{
\item{template}{The location of the template to copy}
}
\description{
Initialise the current directory with skeleton directories and files
to start a new SSR.
}
\details{
*** NB: use a double backslash to escape spaces in directories; e.g.,
        create_SSR_structure(".../01\\ Projects/...")
}
