% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_testresult_string.R
\name{create_testresult_string}
\alias{create_testresult_string}
\title{Return possible test results (repeated symbols)}
\usage{
create_testresult_string(these.times = 0, symbol = "+")
}
\arguments{
\item{these.times}{Number of times to repeat a symbol}

\item{symbol}{Symbol to repeat (default="+")}
}
\value{
the resulting test result string
}
\description{
Given a symbol and a number (x), repeat that symbol x
number of times.  Used to generate test results (e.g.,
for protein dipstick results). The number (x) can be
a vector, to all the generation of a list of possible
testresult strings.
}
\examples{
create_testresult_string( 5 )
create_testresult_string( 1:3, "-" )
}
