% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do.R
\name{do}
\alias{do}
\title{Utility/wrapper function to DO the default things}
\usage{
do()
}
\description{
Simply sources any script called "DO.R" in the current
directory.  Reduces typing for lazy people like me.
}
