% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_count_comparison.R
\name{do_count_comparison}
\alias{do_count_comparison}
\title{Tabulate two separate Yes/No counts and merge together.}
\usage{
do_count_comparison(
  this_data,
  group1_variables,
  group1_column = "Yes",
  group1_name = "Group1",
  group2_variables,
  group2_column = "Yes",
  group2_name = "Group2",
  join_col = "variable",
  this_function = NA,
  key_for_variables = glossary.key
)
}
\arguments{
\item{this_data}{The tibble in which the column exists.}

\item{group1_variables}{The variable to use for the first group.}

\item{group1_column}{The column of interest for the first group (default="Yes").}

\item{group1_name}{The name of the first group (default="Group1").}

\item{group2_variables}{The variable to use for the second group.}

\item{group2_column}{The column of interest for the second group (default="Yes").}

\item{group2_name}{The name of the second group (default="Group1").}

\item{join_col}{The column to use to merge (default="variable").}

\item{this_function}{A label for error/warning messages (by default it is the name of this function).}

\item{key_for_variables}{Glossary to use for parameters.}
}
\value{
The merged summarised counts.
}
\description{
Perform two separate tabulation of Yes/No entries in one or more columns,
for two separate groups.  Then merge these groups based on a specific
variable.  Use this where Yes/No counts for the same measurements are
done at different timepoints (e.g., planned invasive surgery and actual
invasive surgery).
}
