% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_list_comparison.R
\name{do_list_comparison}
\alias{do_list_comparison}
\title{Tabulate two separate list occurence counts and merge together.}
\usage{
do_list_comparison(
  this_data,
  group1_variable,
  group1_column = "count",
  group1_name = "Group1",
  group2_variable,
  group2_column = "count",
  group2_name = "Group2",
  join_col = "variable",
  replacement_value = "Missing",
  this_function = NA,
  vocab_for_variables
)
}
\arguments{
\item{this_data}{The tibble in which the column exists.}

\item{group1_variable}{The variable to use for the first group.}

\item{group1_column}{The column of interest for the first group (default="count").}

\item{group1_name}{The name of the first group (default="Group1").}

\item{group2_variable}{The variable to use for the second group.}

\item{group2_column}{The column of interest for the second group (default="count").}

\item{group2_name}{The name of the second group (default="Group2").}

\item{join_col}{The column to use to merge (default="variable").}

\item{replacement_value}{What to replace empty fields with (default="Missing").}

\item{this_function}{A label for error/warning messages (by default it is the name of this function).}

\item{vocab_for_variables}{The vocabulary for the parameters, as extracted via parse_MACRO_DSD_file()}
}
\value{
The merged summarised counts.
}
\description{
Perform counts of occurences in a list for two separate groups.
Then merge these groups based on a specific variable.
}
