% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_list_extraction.R
\name{do_list_extraction}
\alias{do_list_extraction}
\title{Count the occurence of items in a list}
\usage{
do_list_extraction(
  this_data,
  this_var,
  add_total = TRUE,
  expand_levels = TRUE,
  order_by_vocab = FALSE,
  missing_data_limit = 10,
  this_function = NA,
  vocab_for_variables
)
}
\arguments{
\item{this_data}{The tibble in which the column exists.}

\item{this_var}{The column name (if not a quosure, it will be made into one).}

\item{add_total}{If true, add a line to the table with the totals for all numeric columns (default=TRUE).}

\item{expand_levels}{If true, include all levels of the factor, as defined in the glossary (i.e., as extracted from the DSD).  Default = TRUE.}

\item{order_by_vocab}{If true, present levels of the factor in the order defined parameter vocabulary (i.e., as extracted from the DSD).  Default = FALSE}

\item{missing_data_limit}{The limit for missing data, beyond which a warning is thrown.}

\item{this_function}{A label for error/warning messages (by default it is the name of this function).}

\item{vocab_for_variables}{The vocabulary for the parameters, as extracted via parse_MACRO_DSD_file()}
}
\value{
Counts of levels in the factor.
}
\description{
Summarises the items in a list and provides a count of each.
Missing values are replaced with "Missing".
}
