% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_summary.R
\name{do_summary}
\alias{do_summary}
\title{Summarise numerical data into a table and boxplot}
\usage{
do_summary(
  this_data,
  this_var,
  col_limit = NULL,
  scale_string = "free",
  missing_data_limit = 10,
  this_function = NA,
  key_for_variables = glossary.key
)
}
\arguments{
\item{this_data}{The tibble in which the column exists}

\item{this_var}{The column name (if not a quosure, it will be made into one)}

\item{col_limit}{Limit the number of graphs per row in a facet_wrap}

\item{scale_string}{Whether scales should be free, free_y, free_x etc (default=free)}

\item{missing_data_limit}{Print warning for any result showing at least this percentage of missing data.}

\item{this_function}{A label for error/warning messages (by default it is the name of this function).}

\item{key_for_variables}{Glossary to use for parameters}
}
\value{
A list containing (1) table (2) plot
}
\description{
This functions takes one or more numerical variables
(as quo or quos respectively) and summarises the data
as a table and as a plot.  Both are returned in a list.
}
