% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_definitions.R
\docType{data}
\name{example_trial.data}
\alias{example_trial.data}
\title{Example Trial Data}
\format{
An object of class \code{"tibble"}
}
\usage{
data(example_trial_data)
}
\description{
Completely random data generated (using simstudy) to represent
a clinical trial.  Proportions for ethnicity informed by the 
Scottish 2011 census (https://www.scotlandscensus.gov.uk/variables-classification/ethnic-group).
Surgical outcome hierarchy taken from NAXIVA (https://clinicaltrials.gov/ct2/show/NCT03494816).
}
\examples{
data(example_trial_data)
age = example_trial.data$age
\donttest{boxplot(age)}
}
\keyword{datasets}
