% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_definitions.R
\docType{data}
\name{example_trial.glossary}
\alias{example_trial.glossary}
\title{Glossary for example Trial Data}
\format{
A named list
}
\usage{
data(example_trial_data)
}
\description{
A named list containing a more readable string for each variable in the dataset
}
\examples{
data(example_trial_data)
age = example_trial.glossary$age
}
\keyword{datasets}
