% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_definitions.R
\docType{data}
\name{example_trial.vocabulary}
\alias{example_trial.vocabulary}
\title{Vocabulary for example Trial Data}
\format{
An object of class \code{"tibble"}
}
\usage{
data(example_trial_data)
}
\description{
A named list containing a list of acceptable values for each variable in the dataset
}
\examples{
data(example_trial_data)
age = example_trial.vocabulary$Gender
}
\keyword{datasets}
