% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_dependency_locations_in_glossary.R
\name{expand_dependency_locations_in_glossary}
\alias{expand_dependency_locations_in_glossary}
\title{Expand dependency locations in glossary}
\usage{
expand_dependency_locations_in_glossary(glossary)
}
\arguments{
\item{glossary}{A glossary object (likely generated by parse_MACRO_DSD_file())}
}
\value{
The glossary with dependencies updated.
}
\description{
Not all dependences in the Definitions Study Document
are fully defined - some are missing their visit/section
labels.  To link these dependencies properly, we need to
fill in the missing information.  This function takes an
existing glossary object and adds the additional dependency
information.
}
