% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_glossary.R
\name{expand_glossary}
\alias{expand_glossary}
\title{Expand glossary}
\usage{
expand_glossary(glossary.in, expansion_list, verbose, silent)
}
\arguments{
\item{glossary.in}{The existing glossary object.}

\item{expansion_list}{Which test batteries should be added to which visits.}

\item{verbose}{Set to true/false to override the default (VERBOSE_REPORTING).}

\item{silent}{Set to true/false to override the default (SILENT_REPORTING).}
}
\value{
An updated glossary object with new test batteries added.
}
\description{
The DSD document will only define the fields for a
test battery for the first visit during which that test
battery is used.  This function will expand the glossary
to include all visit that use a test battery, as specificed
by the expansion list provided.
}
