% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dummy_data.R
\name{expand_trial_data}
\alias{expand_trial_data}
\title{Expand trial data (internal)}
\usage{
expand_trial_data(d.0, d.toexpand, column_name = "new_column", option_list)
}
\arguments{
\item{d.0}{Data in}

\item{d.toexpand}{Data to expand}

\item{column_name}{This column}

\item{option_list}{Choose random data from these options}
}
\value{
New dataset
}
\description{
Expand trial data (internal)
}
