% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_complex_dependencies.R
\name{extract_complex_dependencies}
\alias{extract_complex_dependencies}
\title{Extract complex dependencies.}
\usage{
extract_complex_dependencies(
  this_data,
  this_glossary,
  id_var,
  site_var,
  these_dependencies
)
}
\arguments{
\item{this_data}{The tibble to be checked for missing data.}

\item{this_glossary}{The glossary dictionary with parameter visit/test battery etc information.}

\item{id_var}{A quosure representing the column that contains the patient label.}

\item{site_var}{A quosure representing the column that contains the site name for that patient.}

\item{these_dependencies}{A dependencies object (likely generated by parse_MACRO_DSD_file_for_glossary_and_dependencies()).}
}
\value{
Missing data percentages in a tibble
}
\description{
Extract complex dependencies.
}
