% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_mode.R
\name{find_mode}
\alias{find_mode}
\title{Find the mode in the list of values (i.e., the most frequent value)}
\usage{
find_mode(v)
}
\arguments{
\item{v}{The list of values for which to find the mode}
}
\value{
The mode
}
\description{
Find the mode in the list of values (i.e., the most frequent value)
}
