% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_flextable_to_page.R
\name{fit_flextable_to_page}
\alias{fit_flextable_to_page}
\title{Fit flextable object to the width of the page}
\usage{
fit_flextable_to_page(ft, pgwidth = 6)
}
\arguments{
\item{ft}{The flextable object}

\item{pgwidth}{The width of the page, in inches (default = 6in)}
}
\value{
The flextable with the width updated
}
\description{
Fit flextable object to the width of the page
}
