% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_visit_label.R
\name{generate_visit_label}
\alias{generate_visit_label}
\title{Generate a label for each visit for display purposes}
\usage{
generate_visit_label(t, s, v)
}
\arguments{
\item{t}{The string representing the visit parameter name (original column name)}

\item{s}{The string to remove from the visit parameter name}

\item{v}{The string containing the overall visit name}
}
\value{
The label string for the visit
}
\description{
Generate a label for each visit for display purposes
}
