% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dummy_data.R
\name{get_random_dates}
\alias{get_random_dates}
\title{Generate a random sample of dates in a given range}
\usage{
get_random_dates(
  number_of_patients = 100,
  from = today() - months(3),
  to = today() + months(3)
)
}
\arguments{
\item{number_of_patients}{The number of patients in the dataset (default=100)}

\item{from}{Start date for random date range (default is 3 months ago)}

\item{to}{End date for random date range (default is 3 months in the future)}
}
\value{
a list of random dates in the given range
}
\description{
Generate a random sample of dates in a given range
}
\examples{
get_random_dates(10)
get_random_dates(100,as.Date("2000-01-01"),as.Date( "2000-12-31"))
}
