% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialise_RECIST_VALUES.R
\name{initialise_RECIST_VALUES}
\alias{initialise_RECIST_VALUES}
\title{Initialise RECIST values as global variables}
\usage{
initialise_RECIST_VALUES(version = "1.1")
}
\arguments{
\item{version}{The version of RECIST to be used (currently only 1.1)}
}
\value{
NA
}
\description{
Initialise RECIST values as global variables
}
