% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/parse_MACRO_DSD_for_glossary_and_dependencies2.R
\name{parse_MACRO_DSD_file_for_glossary_and_dependencies}
\alias{parse_MACRO_DSD_file_for_glossary_and_dependencies}
\title{Parse information from DSD (MACRO) document}
\usage{
parse_MACRO_DSD_file_for_glossary_and_dependencies(
  XML_DSD_LOCATION,
  verbose,
  silent
)
}
\arguments{
\item{XML_DSD_LOCATION}{The location of the DSD file as extracted from MACRO}

\item{verbose}{Set to true/false to override the default (VERBOSE_REPORTING)}

\item{silent}{Set to true/false to override the default (SILENT_REPORTING)}
}
\value{
A glossary key object.
}
\description{
This function extracts the relevant information from a
Definition Study Document (DSD), obtained via MACRO, that
describes the visits, sections and fields that comprise
the CRF for a study.  It builds a glossary key with the
information that it extracts to allow informative annotation
of fields in SSRs.
}
