% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_attendance_data.R
\name{plot_attendance_data}
\alias{plot_attendance_data}
\title{CAlculate how much data is missing for each patient, visit and test battery.}
\usage{
plot_attendance_data(
  missing_data_percentages,
  id_var,
  site_var,
  patient_attendance,
  width.expansion = 15,
  height.expansion = 8,
  first_column.width = unit(4, "cm"),
  visit_string_removal = "Date_Form_Comp",
  panel_sizes = c(1, 1)
)
}
\arguments{
\item{missing_data_percentages}{The tibble generated by calculate_missing_data.}

\item{id_var}{A quosure representing the column that contains the patient label.}

\item{site_var}{A quosure representing the column that contains the site name for that patient.}

\item{patient_attendance}{The tibble representing attendance at visits (containing the patient label, hospital, visit annotation and all dates).}

\item{width.expansion}{A coefficient to control the width of the panels for each site}

\item{height.expansion}{A coefficient to control the height of the panels for each visit}

\item{first_column.width}{A measurement (using unit()) for the width of the row labels}

\item{visit_string_removal}{A string to remove from the visit column name for display}

\item{panel_sizes}{The relative sizes (heights) of the two panels}
}
\description{
CAlculate how much data is missing for each patient, visit and test battery.
}
