% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_missing_data.R
\name{plot_missing_data}
\alias{plot_missing_data}
\title{Calculate how much data is missing for each patient, visit and test battery.}
\usage{
plot_missing_data(
  missing_data_percentages,
  id_var,
  site_var,
  patient_annotations,
  annotation_threshold = -1,
  width.expansion = 15,
  first_column.width = unit(4, "cm"),
  panel_sizes = c(1, 1, 1)
)
}
\arguments{
\item{missing_data_percentages}{The tibble generated by calculate_missing_data.}

\item{id_var}{A quosure representing the column that contains the patient label.}

\item{site_var}{A quosure representing the column that contains the site name for that patient.}

\item{patient_annotations}{The tibble representing attendance at visits (containing the patient label, hospital, visit annotation and all dates).}

\item{annotation_threshold}{The threshold of percentage missing data (expressed as a number between 0 and 100) for annotation.  The default is -1 (therefore showing all percentages).}

\item{width.expansion}{A coefficient to control the width of the panels for each site}

\item{first_column.width}{A measurement (using unit()) for the width of the row labels}

\item{panel_sizes}{The relative sizes (heights) of the two panels}
}
\description{
Calculate how much data is missing for each patient, visit and test battery.
}
