% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_missing_data_detailed.R
\name{plot_missing_data_detailed}
\alias{plot_missing_data_detailed}
\title{Plot the missing data, for specified test batteries and patients}
\usage{
plot_missing_data_detailed(
  missing_data_percentages,
  id_var,
  this_patient,
  this_visit
)
}
\arguments{
\item{missing_data_percentages}{The tibble generated by calculate_missing_data.}

\item{id_var}{A quosure representing the column that contains the patient label.}

\item{this_patient}{The string representing the patient of interest (in the column defined by id_var)}

\item{this_visit}{The string representing the visit of interest}
}
\description{
Plot the missing data, for specified test batteries and patients
}
