% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_data_queries_message.R
\name{print_data_queries_message}
\alias{print_data_queries_message}
\title{Print data queries summary message.}
\usage{
print_data_queries_message(dq.df)
}
\arguments{
\item{dq.df}{The data frame containing all data queries.}
}
\description{
Print data queries summary message.
}
