% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_error.R
\name{print_error}
\alias{print_error}
\title{Print an error message}
\usage{
print_error(message)
}
\arguments{
\item{message}{The error message that is to be printed.}
}
\description{
Print an error to standard out to update the user.
}
\examples{
print_error( "example" )
}
