% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_file_loading_info.R
\name{print_file_loading_info}
\alias{print_file_loading_info}
\title{Print information about which files were loaded.}
\usage{
print_file_loading_info(file_list)
}
\arguments{
\item{file_list}{A list of files that were loaded.}
}
\description{
Print information about which files were loaded.
}
