% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_flextable.R
\name{print_flextable}
\alias{print_flextable}
\title{Print the labetests data}
\usage{
print_flextable(
  tb,
  column_recode = NA,
  bg_colour = "lightgrey",
  column_padding = NA
)
}
\arguments{
\item{tb}{the data as a tibble}

\item{column_recode}{optional named list for renaming columns}

\item{bg_colour}{optional background colour definition for header}

\item{column_padding}{(optional) list of paddings to use for each row; add a new item to the list for each column.}
}
\value{
Table object to print
}
\description{
This function receives data for labtests and formats it
appropriately.  It prints the table as a kable if FLAG.flextable
if FALSE or as a flextable object if FLAG.flextable is TRUE.
}
