% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_labtests_flextable.R
\name{print_labtests_flextable}
\alias{print_labtests_flextable}
\title{Print the labetests data}
\usage{
print_labtests_flextable(
  tb,
  column_recode = NA,
  std_border = fp_border(color = "black", width = 1),
  heavy_border = fp_border(color = "black", width = 2),
  bg_colour = "lightgrey"
)
}
\arguments{
\item{tb}{the data as a tibble}

\item{column_recode}{optional named list for renaming columns}

\item{std_border}{optional standard border definition for header}

\item{heavy_border}{optional heavy border definition for header}

\item{bg_colour}{optional background colour definition for header}
}
\value{
Table object to print
}
\description{
This function receives data for labtests and formats it
appropriately.  It prints the table as a kable if FLAG.flextable
if FALSE or as a flextable object if FLAG.flextable is TRUE.
}
