% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_message.R
\name{print_message}
\alias{print_message}
\title{Print a message}
\usage{
print_message(
  message,
  print_title = TRUE,
  mask_character = " ",
  section = "CTutils",
  level = 0
)
}
\arguments{
\item{message}{The message that is to be printed.}

\item{print_title}{Print a title to the message (defined in section)}

\item{mask_character}{Use this character in place of the title (if print_title is FALSE).  Default is a space.}

\item{section}{The section name to be printed in message preamble.}

\item{level}{The level of indentation required for this message.}
}
\description{
Print pretty messages to standard out to update the user.
}
\examples{
print_message( "example" )
}
