% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_signsandsyptoms_flextable.R
\name{print_signsandsyptoms_flextable}
\alias{print_signsandsyptoms_flextable}
\title{Print the signs and symptoms data}
\usage{
print_signsandsyptoms_flextable(
  tb,
  column_recode = NA,
  std_border = fp_border(color = "black", width = 1),
  heavy_border = fp_border(color = "black", width = 2),
  bg_colour = "lightgrey",
  highlight_colour.level1 = "#F9E79F",
  highlight_colour.level2 = "#F5B7B1"
)
}
\arguments{
\item{tb}{the data as a tibble}

\item{column_recode}{optional named list for renaming columns}

\item{std_border}{optional standard border definition for header (flextable only)}

\item{heavy_border}{optional heavy border definition for header (flextable only)}

\item{bg_colour}{optional background colour definition for header (flextable only)}

\item{highlight_colour.level1}{optional background colour definition for highlight level 1 (flextable only)}

\item{highlight_colour.level2}{optional background colour definition for highlight level 2 (flextable only)}
}
\value{
Table object to print
}
\description{
This function receives data for signs and symptoms (toxicities)
and formats it appropriately.  It prints the table as a kable if
FLAG.flextable if FALSE or as a flextable object if FLAG.flextable
is TRUE.
}
