% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_EudraCT_vocabulary.R
\name{read_EudraCT_vocabulary}
\alias{read_EudraCT_vocabulary}
\title{Provide necessary EudraCT and RMS vocabulary for use when generating
the EudraCT export}
\usage{
read_EudraCT_vocabulary()
}
\description{
Provide necessary EudraCT and RMS vocabulary for use when generating
the EudraCT export
}
