% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_report_YAML.R
\name{read_report_YAML}
\alias{read_report_YAML}
\title{Read in values from YAML and load them into the global environment}
\usage{
read_report_YAML(yml.file = "settings.yml")
}
\arguments{
\item{yml.file}{The file from which to read}
}
\description{
Read in values from YAML and load them into the global environment
}
