% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RECIST_COMPLETE.R
\name{RECIST_COMPLETE}
\alias{RECIST_COMPLETE}
\title{RECIST_COMPLETE}
\usage{
RECIST_COMPLETE(diam1)
}
\arguments{
\item{diam1}{The individual lesion diameters values at the timepoint under consideration}
}
\value{
TRUE/FALSE indicating whether the data meet the complete criteria
}
\description{
RECIST_COMPLETE
}
