% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RECIST_PARTIAL.R
\name{RECIST_PARTIAL}
\alias{RECIST_PARTIAL}
\title{RECIST_PARTIAL}
\usage{
RECIST_PARTIAL(sum0, sum1)
}
\arguments{
\item{sum0}{The sum values at baseline}

\item{sum1}{The sum values at the timepoint under consideration}
}
\value{
TRUE/FALSE indicating whether the data meet the partial criteria
}
\description{
RECIST_PARTIAL
}
