% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RECIST_PARTIAL.R
\name{RECIST_PARTIAL_calculation}
\alias{RECIST_PARTIAL_calculation}
\title{RECIST_PARTIAL calculation}
\usage{
RECIST_PARTIAL_calculation(v0, v1)
}
\arguments{
\item{v0}{The sum values at baseline}

\item{v1}{The sum values at the timepoint under consideration}
}
\value{
the calculated value that is relevant to the PARTIAL RECIST assessment
}
\description{
RECIST_PARTIAL calculation
}
