% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RECIST_PROGRESSIVE.R
\name{RECIST_PROGRESSIVE}
\alias{RECIST_PROGRESSIVE}
\title{RECIST_PROGRESSIVE}
\usage{
RECIST_PROGRESSIVE(sum0, sum1, diam0, diam1)
}
\arguments{
\item{sum0}{The sum values at baseline}

\item{sum1}{The sum values at the timepoint under consideration}

\item{diam0}{The diameter values at baseline}

\item{diam1}{The diameter values at the timepoint under consideration}
}
\value{
TRUE/FALSE indicating whether the data meet the progressive criteria
}
\description{
RECIST_PROGRESSIVE
}
