% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RECIST_PROGRESSIVE.R
\name{RECIST_PROGRESSIVEabs_calculation}
\alias{RECIST_PROGRESSIVEabs_calculation}
\title{RECIST PROGRESSIVE absolute increase calculation}
\usage{
RECIST_PROGRESSIVEabs_calculation(v0, v1)
}
\arguments{
\item{v0}{The sum values at baseline}

\item{v1}{The sum values at the timepoint under consideration}
}
\value{
the calculated value that is relevant to the PROGRESSIVE RECIST assessment (absolute increase)
}
\description{
RECIST PROGRESSIVE absolute increase calculation
}
