% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RECIST_PROGRESSIVE.R
\name{RECIST_PROGRESSIVEperc_calculation}
\alias{RECIST_PROGRESSIVEperc_calculation}
\title{RECIST PROGRESSIVE percentage increase calculation}
\usage{
RECIST_PROGRESSIVEperc_calculation(v0, v1)
}
\arguments{
\item{v0}{The sum values at baseline}

\item{v1}{The sum values at the timepoint under consideration}
}
\value{
the calculated value that is relevant to the PROGRESSIVE RECIST assessment (percentage increase)
}
\description{
RECIST PROGRESSIVE percentage increase calculation
}
