% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/record_data_query.R
\name{record_data_query}
\alias{record_data_query}
\title{Record a data query}
\usage{
record_data_query(
  patient_list,
  query_message,
  report_reference = NULL,
  CRF_reference = NULL,
  list_of_data_queries
)
}
\arguments{
\item{patient_list}{A list of patients}

\item{query_message}{The query message of each patient}

\item{report_reference}{A reference for the DMC report}

\item{CRF_reference}{A reference for the CRF}

\item{list_of_data_queries}{A list of data queries to add to}
}
\value{
An updated list of data queries
}
\description{
data_queries.list = NULL
Add a data query to the list of queries being compiled throughout
the report writing process.
}
