% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refactor_variables.R
\name{refactor_variable}
\alias{refactor_variable}
\title{Change the levels of a factor to use an explicit list}
\usage{
refactor_variable(
  this.data,
  this.var,
  these.levels = c("Yes", "No", "Missing")
)
}
\arguments{
\item{this.data}{The tibble in which the column exists}

\item{this.var}{The column name to refactor}

\item{these.levels}{The new levels for this factor (default = Yes, No, Missing)}
}
\value{
The updated tibble
}
\description{
Change the levels of a factor to use an explicit list
}
