% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rejig_panel_dimensions.R
\name{rejig_panel_dimensions}
\alias{rejig_panel_dimensions}
\title{Rejig panel dimensions}
\usage{
rejig_panel_dimensions(
  p,
  w,
  h,
  w.expansion = 15,
  h.expansion = 8,
  c1.w = unit(4, "cm")
)
}
\arguments{
\item{p}{The plot to rejig}

\item{w}{The relative widths to use}

\item{h}{The relative heights to use}

\item{w.expansion}{A value to expand the widths by}

\item{h.expansion}{A value to expand the heights by}

\item{c1.w}{The width of the first column}
}
\value{
Corrected plot
}
\description{
Rejig panel dimensions
}
