% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rejig_panel_widths.R
\name{rejig_panel_widths}
\alias{rejig_panel_widths}
\title{Rejig panel widths}
\usage{
rejig_panel_widths(p, w, w.expansion = 15, c1.w = unit(4, "cm"))
}
\arguments{
\item{p}{The plot to rejig}

\item{w}{The relative widths to use}

\item{w.expansion}{A value to expand the widths by}

\item{c1.w}{The width of the first column}
}
\value{
Corrected plot
}
\description{
Rejig panel widths
}
