% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_func.R
\name{replace_func}
\alias{replace_func}
\title{Replace NAs with another value}
\usage{
replace_func(val, replace)
}
\arguments{
\item{val}{The values in which to look for NAs}

\item{replace}{The value to replace the NAs in val}
}
\value{
The values with NAs replaced
}
\description{
Replace NAs with another value
}
