% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_infinite_numbers.R
\name{replace_infinite_numbers}
\alias{replace_infinite_numbers}
\title{Replace infinite numbers with NA}
\usage{
replace_infinite_numbers(x)
}
\arguments{
\item{x}{the values to be checked}
}
\value{
the modified list of values
}
\description{
Replace infinite numbers with NA
}
