% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_missing_number.R
\name{replace_missing_number}
\alias{replace_missing_number}
\title{Replace missing numeric value with another number.}
\usage{
replace_missing_number(
  this_data,
  this_var,
  this_function = NA,
  replacement_targets = NA,
  replacement_value = 0
)
}
\arguments{
\item{this_data}{The tibble in which the column exists}

\item{this_var}{The column name (if not a quosure, it will be made into one)}

\item{this_function}{A label for error/warning messages (by default it is the name of this function).}

\item{replacement_targets}{The value to be replaced (default=NA)}

\item{replacement_value}{The value to replace it with (default=0)}
}
\value{
The updated tibble
}
\description{
Given a tibble and a column, this function will find target
values (by default NA) and replace them with another number (by
default 0).
}
