% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_missing_string.R
\name{replace_missing_string}
\alias{replace_missing_string}
\title{Replace missing values with another string.}
\usage{
replace_missing_string(
  this_data,
  this_var,
  this_function = NA,
  replacement_targets = "",
  replacement_value = "Missing"
)
}
\arguments{
\item{this_data}{The tibble in which the column exists}

\item{this_var}{The column name (if not a quosure, it will be made into one)}

\item{this_function}{A label for error/warning messages (by default it is the name of this function).}

\item{replacement_targets}{The value to be replace (default="")}

\item{replacement_value}{The value to replace it with (default="Missing)}
}
\value{
The updated tibble
}
\description{
Given a tibble and a column, this function will find target
values (by default "") and replace them with a string (by
default "Missing").
}
\details{
Function can be repurposed to replace any target value
with another string.
}
