% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolve_individual_filter_statement.R
\name{resolve_individual_filter_statement}
\alias{resolve_individual_filter_statement}
\title{Resolves a dependency, so that it can be used as a filter statement.}
\usage{
resolve_individual_filter_statement(s, c, g, dm)
}
\arguments{
\item{s}{The dependency (character string).}

\item{c}{The column to which this dependency belongs.}

\item{g}{The glossary in which we can check that fields exist.}

\item{dm}{The dependency map (contains vocabulary in the form that we need).}
}
\value{
The resolved dependency.
}
\description{
Receives a dependency (s) and splits it into a list of its constituent
components.  Each component is then checked:
1. Does it check against a value (i.e. does it contain a =?)
   a) is it a does not equal (!=) rather than an equal (=)?
   b) does the field exist in the glossary?
   c) does the value need to be translated using the vocabulary?
}
