% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_image.R
\name{save_image}
\alias{save_image}
\title{Save CT data to file.}
\usage{
save_image(infile, verbose = TRUE)
}
\arguments{
\item{infile}{Path to the input file}

\item{verbose}{Be verbose?  Default is true.}
}
\description{
This function saves the current version of the data to file.
It should be run between every chapter of the report to ensure
that all data are available.  If the infile parameter is NULL, the
globally defined location for the data extract file will be used.
}
