% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subchunkify.R
\name{subchunkify}
\alias{subchunkify}
\title{Print a figure at a different width/height than the chunk
it is in.}
\usage{
subchunkify(g, fig_height = 7, fig_width = 5)
}
\arguments{
\item{g}{The plot}

\item{fig_height}{The desired height of the plot (inches)}

\item{fig_width}{The desired width of the plot (inches)}
}
\description{
Print a figure at a different width/height than the chunk
it is in.
}
