% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_fn.R
\name{test_fn}
\alias{test_fn}
\title{Print a name}
\usage{
test_fn(data_in)
}
\arguments{
\item{data_in}{Name to be printed}
}
\value{
The string printed
}
\description{
This function takes a name and prints it.
}
