% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_logical_expression.R
\name{tidy_logical_expression}
\alias{tidy_logical_expression}
\title{Tidy/reformat a dependency}
\usage{
tidy_logical_expression(t)
}
\arguments{
\item{t}{The dependency (character string).}
}
\value{
The tidied dependency.
}
\description{
Tidy and reformat a dependency between fields so that it can be used
in a filter statement. Steps are:
1. remove repeated ands/ors (these might occur if a dependency was removed earlier)
2. remove leading/training/internal space
3. remove ands/ors at the beginning/end of the string
4. replace and/or with &/|
5. replace = with ==
6. (but change !== back to !=)
}
