% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTutilsproject.R
\name{CTutilsproject}
\alias{CTutilsproject}
\title{CTutilsproject}
\usage{
CTutilsproject(
  path,
  author_name = "Lisa Hopcroft",
  author_email = "Lisa.Hopcroft@phs.scot",
  author_organisation = "Public Health Scotland",
  trial_name = "TRIAL_NAME",
  n_scripts = 1,
  git = TRUE,
  renv = FALSE
)
}
\arguments{
\item{path}{Filepath for the project.}

\item{author_name}{Name of the main author for the project.}

\item{n_scripts}{Number of code files to start the project with.}

\item{git}{Initialise the project with Git.}

\item{renv}{Initialise the project with package management using renv.}
}
\value{
New project created according to the CTutils project structure.
}
\description{
Create a directory structure to match that used by CTutils (PHS). This function is meant to be used within RStudio by going to the File menu, then New Project.
}
\examples{
\dontrun{
phsproject(path = file.path(getwd(), "testproj"), author_name = "A Person", n_scripts = 1)
}

}
